chore: move types from shared-data to performance-metrics
