+++ performance-metrics/src/performance_metrics/data_definitions.py
@@ -1,0 +1,36 @@
+"""Defines data structures used in the performance metrics module."""
+
+import enum
+
+
+class RobotContextState(enum.Enum):
+    """Enum representing different states of a robot's operation context."""
+
+    STARTING_UP = 0, "STARTING_UP"
+    CALIBRATING = 1, "CALIBRATING"
+    ANALYZING_PROTOCOL = 2, "ANALYZING_PROTOCOL"
+    RUNNING_PROTOCOL = 3, "RUNNING_PROTOCOL"
+    SHUTTING_DOWN = 4, "SHUTTING_DOWN"
+
+    def __init__(self, state_id: int, state_name: str) -> None:
+        """Initialize the enum member."""
+        self.state_id = state_id
+        self.state_name = state_name
+
+    @classmethod
+    def from_id(cls, state_id: int) -> "RobotContextState":
+        """Returns the enum member matching the given state ID.
+
+        Args:
+            state_id: The ID of the state to retrieve.
+
+        Returns:
+            RobotContextStates: The enum member corresponding to the given ID.
+
+        Raises:
+            ValueError: If no matching state is found.
+        """
+        for state in RobotContextState:
+            if state.state_id == state_id:
+                return state
+        raise ValueError(f"Invalid state id: {state_id}")
