+++ performance-metrics/tests/performance_metrics/test_robot_context_tracker.py
@@ -4,7 +4,7 @@ import asyncio
 from pathlib import Path
 import pytest
 from performance_metrics.robot_context_tracker import RobotContextTracker
-from opentrons_shared_data.performance.dev_types import RobotContextState
+from performance_metrics.data_definitions import RobotContextState
 from time import sleep, time_ns
 from unittest.mock import patch
