--- shared-data/python/opentrons_shared_data/performance/__init__.py
@@ -1,1 +1,0 @@
-"""Performance metrics."""
